# cmd/target-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ownership verification and authorization scope gate for the Scantinel security-scanning platform. The POST /internal/v1/scope/check endpoint is the hard gate that scan-service and agent-service MUST consult before dispatching any intrusive task.",
        "title": "target-service API",
        "termsOfService": "https://scantinel.ai/terms",
        "contact": {
            "name": "Scantinel Support",
            "url": "https://scantinel.ai/contact",
            "email": "security@scantinel.ai"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/admin/targets": {
            "get": {
                "description": "Cross-user paged list. Unknown filter values return 400. `q` is a case-insensitive substring match on value or label.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List targets across all users (ADMIN)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Must be \\",
                        "name": "X-Is-Admin",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Filter by owner",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "unverified|verifying|verified|revoked",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "customer|program",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Substring match on value/label",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 50, max 200)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AdminTargetListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "403": {
                        "description": "admin only",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Registers a target as verified with a method=program authorization whose evidence is the public program. user_id defaults to the admin's X-User-Id; authorized_days 1..365 (default 90). Value rules match POST /v1/targets (ip/cidr still refused while IP targets are disabled).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a bug-bounty program target (ADMIN)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Must be \\",
                        "name": "X-Is-Admin",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Admin's platform user id (audit actor, default owner)",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Program target payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AdminCreateProgramTargetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TargetAdminDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "403": {
                        "description": "admin only",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "409": {
                        "description": "duplicate (user_id, kind, value)",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "422": {
                        "description": "IP/CIDR targets are not yet supported",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/v1/admin/targets/{target_uid}": {
            "get": {
                "description": "Target detail with its current authorization (incl. evidence) and asset inventory, regardless of owner.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get any user's target (ADMIN)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Must be \\",
                        "name": "X-Is-Admin",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target UID",
                        "name": "target_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TargetAdminDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "403": {
                        "description": "admin only",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes the target with its authorizations and asset inventory regardless of owner.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete any user's target (ADMIN)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Must be \\",
                        "name": "X-Is-Admin",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Admin's platform user id (audit actor)",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target UID",
                        "name": "target_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "403": {
                        "description": "admin only",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/targets": {
            "get": {
                "description": "Return a paged list of targets owned by the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "List targets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authenticated user ID",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 10, max 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.TargetDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Register a domain/URL/IP/CIDR target and receive a verification challenge token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "Register a new target",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authenticated user ID (stamped by Traefik)",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Target registration payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateTargetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TargetDetailDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "409": {
                        "description": "duplicate (kind, value) for this user",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "422": {
                        "description": "IP/CIDR targets are not yet supported",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/v1/targets/{target_uid}": {
            "get": {
                "description": "Return a target with its current authorization scope and asset inventory.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "Get target detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authenticated user ID",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target UID",
                        "name": "target_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TargetDetailDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a target and all its authorization scope and asset inventory.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "targets"
                ],
                "summary": "Delete a target",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authenticated user ID",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target UID",
                        "name": "target_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AdminCreateProgramTargetRequest": {
            "type": "object",
            "required": [
                "kind",
                "program",
                "value"
            ],
            "properties": {
                "kind": {
                    "type": "string",
                    "enum": [
                        "domain",
                        "url",
                        "ip",
                        "cidr"
                    ]
                },
                "label": {
                    "type": "string",
                    "maxLength": 255
                },
                "program": {
                    "$ref": "#/definitions/model.ProgramEvidence"
                },
                "user_id": {
                    "description": "defaults to the admin's own X-User-Id",
                    "type": "integer",
                    "minimum": 1
                },
                "value": {
                    "type": "string",
                    "maxLength": 512,
                    "minLength": 1
                }
            }
        },
...
        "model.AdminTargetListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TargetAdminDTO"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" | \"error\"",
                    "type": "string"
                }
            }
        },
...
        "model.TargetAdminDTO": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AssetDTO"
                    }
                },
                "authorization": {
                    "$ref": "#/definitions/model.AuthorizationDTO"
                },
                "created_at": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "registrable_domain": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    }
}